+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - 2.0.0
     - Hash function is now configurable
+    - Can now iterate over values in Boolean Array
     - Changed name of num slices argument
     - Method `numHashes()` is now `numSlices()`